backend-py/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
sqlalchemy>=2.0
pydantic>=2.5
pydantic-settings>=2.0
psycopg2-binary>=2.9
azure-ai-inference>=1.0.0b7
langchain>=0.2
langchain-community>=0.2
langchain-core>=0.2
langchain-ollama>=0.1
pyodbc>=5.1.0
pgvector>=0.3.6
kafka-python>=2.0.2

# CORS and utilities
python-multipart==0.0.6
aiohttp==3.9.0
